+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### Maintenance
 * Updated `Notification` component. Do not show browser notifications when window is active
+* Fixed `Select`'s `list` slot. Didn't work before
+* Added `selected` key to `Select`'s `options` object. Useful for `option` slot
+
+
 
 
 ## 0.2.11